TASK: Convert Drawer/MenuItemGroup to typescript
